horde

Module collect

An API for quiescent state based reclamation.

Structs

Pin

Represents a proof that no deferred methods will run for the lifetime 'a. It can be used to access data structures in a lock-free manner.

Functions

collect

Signals a quiescent state where garbage may be collected.

defer_unchecked⚠

This schedules a closure to run at some point after all threads are outside their current pinned regions.

pin

Marks the current thread as pinned and returns a proof of that to the closure.

release

Removes the current thread from the threads allowed to access lock-free data structures.