horde::collect

Function release

pub fn release()

Removes the current thread from the threads allowed to access lock-free data structures.

This allows memory to be freed without waiting for collect calls from the current thread. pin can be called after to continue accessing lock-free data structures.

This will not free any garbage so collect should be called before the last thread terminates to avoid memory leaks.

Calling this function when the thread is already unregistered is a no-op.

This will panic if called while the current thread is pinned or during a deferred callback.