horde::sync_push_vec

Struct SyncPushVec

pub struct SyncPushVec<T> { /* private fields */ }

A contiguous push-only array type with lock-free reads.

Implementations

impl<T> SyncPushVec<T>

pub fn new() -> Self

Constructs a new, empty vector with zero capacity.

The vector will not allocate until elements are pushed onto it.

pub fn with_capacity(capacity: usize) -> Self

Constructs a new, empty vector with the specified capacity.

The vector will be able to hold exactly capacity elements without reallocating. If capacity is 0, the vector will not allocate.

pub fn mutex(&self) -> &Mutex<()>

Gets a reference to the underlying mutex that protects writes.

pub fn read<'a>(&'a self, pin: Pin<'a>) -> Read<'a, T>

Creates a Read handle from a pinned region.

Use crate::collect::pin to get a Pin instance.

pub unsafe fn unsafe_write(&self) -> Write<'_, T>

Creates a Write handle without checking for exclusive access.

Safety

It’s up to the caller to ensure only one thread writes to the vector at a time.

pub fn write(&mut self) -> Write<'_, T>

Creates a Write handle from a mutable reference.

pub fn lock(&self) -> LockedWrite<'_, T>

Creates a LockedWrite handle by taking the underlying mutex that protects writes.

pub fn lock_from_guard<'a>( &'a self, guard: MutexGuard<'a, ()>, ) -> LockedWrite<'a, T>

Creates a LockedWrite handle from a guard protecting the underlying mutex that protects writes.

pub fn as_mut_slice(&mut self) -> &mut [T]

Extracts a mutable slice of the entire vector.