Module horde::sync_table

A hash table with lock-free reads.

It is based on the table from the hashbrown crate.

Structs

Iter
An iterator over the entries of a HashMap.

LockedWrite
A handle to a SyncTable with write access protected by a lock.

PotentialSlot
Represents where a value would be if inserted.

Read
A handle to a SyncTable with read access.

SyncTable
A hash table with lock-free reads.

Write
A handle to a SyncTable with write access.

Functions

shard_index_by_hash
Get a suitable shard index from a hash when sharding the hash table.

Type Definitions

DefaultHashBuilder
Default hash builder for SyncTable.